Refactor the PageNotFound component and moved styles to the assets views folder
